apache plugin: Fix a segmentation fault in the config handling of VerifyPeer / VerifyHost.
Hello,

Today I upgraded a machine from 4.5.2 to 4.9.0. Everything went fine, but
the new version segfaults immediately at startup. With some help from gdb
I tracked down the problem to the apache module. It will occur if one
tries to use the VerifyPeer/VerifyHost options (I guess few people use
these as I don't see complaints?)

When the plugin attempts to read the value of a boolean option, instead of
using the pre-parsed int value, it misdetects the type of the option
(seems a mechanical mistake) and tries to use the string pointer, which is
actually NULL.
I'm sending a patch, it's actually shorter than the description ;)

Thanks to all developers and maintainers of collectd for working on it.
Have a nice holidays.