openvpn plugin: Rename the "OnlyAggregateUsers" option to "CollectIndividualUsers".
The semantic is inverted: When true (default) statistics for individual
users will be collected.

The "AggregateUsers" is now named "CollectUserCount" and a new name for
"Compression", "CollectCompression", has been added. This makes the names
of the config options consistent with one another.

The module global configuration variables have been changed to _Bool
while I was at it.