Introduces ExcludeRegexp to the tail plugin
Hi Florian

Florian Forster wrote:
> On Fri, Mar 19, 2010 at 06:13:14PM +0100, Peter Warasin wrote:
>> I'm searching for something like "ExcludeRegExp".
>
> no, something like that doesn't exist yet. I like the idea though, so
> I've added it to the [[Roadmap]] wiki page.

Attached is a patch which implements this.
The patch is against 4.9.1

Please tell me if you want me to rebase it to current svn

kind regards,
peter

--
:: e n d i a n
:: open source - open minds

:: peter warasin
:: http://www.endian.com   :: peter@endian.com

Adds ExcludeRegex to the tail-match plugin

With this keyword it is possible to exclude a line from a match.

Signed-off-by: Florian Forster <octo@huhu.verplant.org>