Bugfix for libcollectdclient lcc_putval.
Hello,

I'm using collectd to monitor system and process parameters of our code under
test running on a cluster. The test processes communicate to collectd via the
unixsock plugin, but there was a small error in the client library which I had
to fix first.

The wire format of the putval command is missing a space behind the closing
double quote of the identifier. As I said, it is a small fix, but nevertheless
I attached the patch for client.c. The bug is present in 4.6 (that is the
version I'm using in production), but I fixed it against 4.10.0.

Cheers,
Johan

Signed-off-by: Florian Forster <octo@huhu.verplant.org>