amqp plugin: Implement persistent connection handling.
A connection is now opened on the first value to be written and is kept
open for as long as possible. This is hopefully much more efficient than
opening a new connection for each value.

The value_list_t is now converted to JSON first, before locking the
global connection object, in order to hold the lock for the least
possible time.