Add collectd-flush command line utility.
collectd-flush is a small command-line utility which uses libcollectdclient
to flush collectd through the unixsock plugin. As far as I know, there is
no easy way to work with unix sockets from the shell, so if someone
wanted to flush collectd without this program, they would need to do it
through some other language (or restart the deamon).

I am no C wizard, so please bear with me if I have
done something a stupid way. :)
The code is based on/inspired by collectd-nagios.

If this message is not properly encoded as utf-8, something wrong must have
happened with my mailer. If that's the case, I'm sorry.  The only non-ascii character
in the patch is the second character of my name, which is an &aring;.

Signed-off-by: Håkon J Dugstad Johnsen <hakon-dugstad.johnsen@telenor.com>
Signed-off-by: Florian Forster <octo@huhu.verplant.org>