Update Makefile.am to build collectd-flush
As I am no automake expert, I basically copied
the lines for collectd-nagios and changed them
to collectd-flush. It works, but maybe there could
be better ways to do it. I don't know. :)

As I said in the previous mail:
If this message is not properly encoded as utf-8, something wrong must have
happened with my mailer. If that's the case, I'm sorry.  The only non-ascii character
in the patch is the second character of my name, which is an &aring;.

Signed-off-by: Håkon J Dugstad Johnsen <hakon-dugstad.johnsen@telenor.com>
Signed-off-by: Florian Forster <octo@huhu.verplant.org>