New plugin - lpar
Hello,

here is a patch against 4.10 adding a new plugin "lpar".

LPAR stands for Logical PARtitions, which is the virtualization solution
for IBM high-end power systems running AIX.

The standard cpu plugin shows cpu usage as a percentage of each cpu
available to the system, but in an LPAR the number of cpus and their
apparent power can vary according to the load of every LPAR sharing the
same hardware and to the policy set by the admin.

This new plugin allows to monitor real (physical) CPU usage of the
virtualized system, as well as some other metrics specific to IBM's
partitioning solution.

Regards,

Aurélien Reynaud

--
Love is like PI - natural, irrational, endless, and very important.

Signed-off-by: Florian Forster <octo@huhu.verplant.org>