Vmware Client SDK plugin
https://gist.github.com/712934

I'm not sure the process to get this merged in, but I figured it would
be useful to some people.

This is against collectd-4.7.4, so it may be a little stale as well.

Let me know what's needed to get this merged officially upstream and
I'll be happy to work on it in my spare time.

Thanks!