Use parse_value in the conntrack plugin and submit the result even if it's zero
Dear collectd,

Please find attached a patch for the conntrack plugin.

The patch is about two things:

1. submit the value even if it is zero (which is a legitimate value according
   to types.db);
2. use parse_value and a value_t instead of directly using a double and atof(3).

The first point was important because it meant that the metric was not created
when the initial value was zero. (It could also lead to holes in your graphs).

The parse_value return value is correctly checked, note that the parsed file
ends with a \n which mean that parse_value always complain when running in
debug/info maybe we should replace it with a \0 before handing the buffer to
parse_value() ?

Thanks

--
Louis Opter

Signed-off-by: Florian Forster <octo@collectd.org>