Makefile adds dependency on $(LIBLTDL), which is valued "-lltdl"
When building on OS X, the collectd build failed, because the collectd target
has a dependency on "-lltdl", which is the value the variable $(LIBLTDL) has.
Obviously, this should be the file name instead. I am however not sure how to
specify the correct filename in this case.

Fixes GitHub issue #54.

Signed-off-by: Florian Forster <octo@collectd.org>