Fixed Solaris compilation and runtime issues
Modified ps_get_cmdline. Since in Solaris, /proc/PID/psinfo is a binary file of fixed form, it will always have the same size. There is no need to play with the buffer length since it's clear from the get-go that it will be the sizeof(psinfo_t).

In ps_read(), the size of cmdline cannot be MAX_ARGS. The program will crash and burn. Set the maximum length to the length of psinfo_t.pr_psargs, namely PRARGSZ, currently set to 80.