Added kstat dump to string function
kstat_value_to_string returns the value of a kstat in char format.
It is useful when we want all the values and don't necessarily know the name of the field (similar to a dump). The stats are returned in 2 char arrays, one for the names of the stats and one for their respective values.