network plugin: Do not forward received notifications.
The network plugin would send out all notifications it got through its
callback, even the ones it received itself and even when forwarding was
disabled.

With this patch the network plugin will only transmit notifications created
locally -- forwarding of notifications is not implemented due to missing
loop-detection. For loop-detection we would have to keep track of the last
timestamp of each notifications, and since each host/plugin/type/-instance
field may be empty, this is not as straight forward as with value lists. If
someone needs this, they will need to invest some work I'm afraid.