src/utils_vl_lookup.[ch]: Add module for efficient lookup of VL matches with wildcards.
For a planned aggregation plugin it is important to find "aggregations"
quickly. In order to aggregate anything, they need to include wildcards.
This module provides this basic lookup service, on top of which an
aggregation plugin can easily be built.

This is kept very modular so the code can (hopefully) be used by other
projects as well, in particular c4 could benefit from this lookup structure
as well.