ntpd plugin: Refactor the peer name generation.
Break it out of the main read-loop and put it into two functions: One for the
reference clocks and one for network hosts. The reference clock code will call
the network host code when the clock name is not available in the internal list
of reference clock names. This eliminates the (thread-unsafe) usage of
inet_ntoa(3).