postgresql plugin: Added "Instance" option to <Database> config block.
This option may be used to specify the plugin instance that should be used
instead of the database name. This allows to query multiple databases of the
same name on the same host (e.g. when running multiple database server
versions in parallel).

This fixes GH #161 suggested by Bryan Varner.