Adding strsplitsep() to src/common.c
The new `strsplitsep()` function tokenizes a string into parts and
stores pointers to the parts in `fields`.

This is a generalization of the already-existing `strsplit()`. Instead
of wiring the delimiter string, it receives the `sep` argument (a string
with the separators) used to tokenize the given string.

To minimize code, the `strsplit()` function now becomes a wrapper to the
more general `strsplitsep()` with a fixed `sep` of `" \t\r\n"`.

NOTE: This patch will be used by subsequent changes in the IPMI plugin.