ipmi plugin: Adding support to override the default sensor naming on a per-sensor basis
The IPMI plugin sets a default naming policy for sensors. This feature
allows the user to override the default naming on a per-sensor basis.

To override the default sensor name simply add the custom name after the
default name in the config, as follows:

    Sensor "default_sensor_name;custom_sensor_name"

Note that the `default_sensor_name` is the default name given by the IPMI
plugin. At the moment, there is no easy way to know it beforehand, other
than by running the IPMI plugin and checking the names it assigns to the
sensors.

In the future, a helper tool should exist to list all of the available
sensors with their default IPMI `default_sensor_name`.