ipmi plugin: Adding 'IgnoreSensorFail' option to avoid removing sensors that fail
Before, the IPMI plugin would remove a sensor when a single read failure
occurred. With this option set to `true`, the sensor won't be removed
after failing (i.e., the IPMI plugin will retry reading from the sensor
the next time it gathers data).

By default, it's set to `false`, leaving the default behavior as it was
before the change (i.e. removing sensors even if they fail one reading).