Use the canonical hostname instead of the usual short hostname.
Since the default value for FQDNLookup setting is true, the GenericJMX
plugin should use the canonical hostname by default as well.Although, the C code does make
a check for whehther FQDNLookup is turned on or not and accordingly sets
the hostname. I am not sure how the global configuration properties can
be accessed in Java plugins. Any pointers would be appreciated, so that
I can provide a proper fix.