zfs_arc: support for zfsonlinux
The plugin is structured for Solaris & FreeBSD's get_kstat_value() &
sysctlbyname() functions, which return a single metric based on its
name. ZFS-on-Linux publishes all the stats together in a file in /proc,
which is quite a different approach.
So the idea is to simulate the Solaris/FreeBSD way of doing by using
llist_search() on a linked-list containing all the key/values from the
/proc file. This prevents having to parse it over and over again for
each metric we request.