disk: Add udev-based device renaming on Linux
On systems with large collections of allocated disk resources, the
kernel provided names can be difficult to use to logically group or
compare collected values.  To grant users the ability to assign custom
instance names for disk devices, collectd can query for a specified udev
attribute.  If the attribute is present, its value is subsituted for the
kernel name when submitting the data points.